+++ pkg/config/config.go
@@ -39,8 +39,6 @@ const (
 	// or higher, or the annotation "kubernetes.io/ingress.class" (deprecated).
 	IngressClass = "apisix"
 
-	ApisixRouteClass = ""
-
 	// IngressNetworkingV1 represents ingress.networking/v1
 	IngressNetworkingV1 = "networking/v1"
 	// IngressNetworkingV1beta1 represents ingress.networking/v1beta1
@@ -98,7 +96,6 @@ type KubernetesConfig struct {
 	ElectionID          string             `json:"election_id" yaml:"election_id"`
 	IngressClass        string             `json:"ingress_class" yaml:"ingress_class"`
 	IngressVersion      string             `json:"ingress_version" yaml:"ingress_version"`
-	ApisixRouteClass    string             `json:"apisix_route_class" yaml:"apisix_route_class"`
 	WatchEndpointSlices bool               `json:"watch_endpoint_slices" yaml:"watch_endpoint_slices"`
 	APIVersion          string             `json:"api_version" yaml:"api_version"`
 	EnableGatewayAPI    bool               `json:"enable_gateway_api" yaml:"enable_gateway_api"`
@@ -140,7 +137,6 @@ func NewDefaultConfig() *Config {
 			ResyncInterval:      types.TimeDuration{Duration: 6 * time.Hour},
 			ElectionID:          IngressAPISIXLeader,
 			IngressClass:        IngressClass,
-			ApisixRouteClass:    ApisixRouteClass,
 			IngressVersion:      IngressNetworkingV1,
 			APIVersion:          DefaultAPIVersion,
 			WatchEndpointSlices: false,
