+++ samples/deploy/crd/v1/ApisixRoute.yaml
@@ -69,6 +69,8 @@ spec:
                 - required: ["http"]
                 - required: ["stream"]
               properties:
+                ingressClassName:
+                  type: string
                 http:
                   type: array
                   minItems: 1
@@ -594,197 +596,6 @@ spec:
                             enum: ["endpoint", "service"]
                           subset:
                             type: string
-                        required:
-                          - serviceName
-                          - servicePort
-            status:
-              type: object
-              properties:
-                conditions:
-                  type: array
-                  items:
-                    type: object
-                    properties:
-                      "type":
-                        type: string
-                      reason:
-                        type: string
-                      status:
-                        type: string
-                      message:
-                        type: string
-                      observedGeneration:
-                        type: integer
-    - name: v2
-      served: true
-      storage: true
-      subresources:
-        status: {}
-      additionalPrinterColumns:
-        - jsonPath: .spec.http[].match.hosts
-          name: Hosts
-          type: string
-          priority: 0
-        - jsonPath: .spec.http[].match.paths
-          name: URIs
-          type: string
-          priority: 0
-        - jsonPath: .spec.http[].backends[].serviceName
-          name: Target Service(HTTP)
-          type: string
-          priority: 1
-        - jsonPath: .spec.tcp[].match.ingressPort
-          name: Ingress Server Port(TCP)
-          type: integer
-          priority: 1
-        - jsonPath: .spec.tcp[].match.backend.serviceName
-          name: Target Service(TCP)
-          type: string
-          priority: 1
-        - jsonPath: .metadata.creationTimestamp
-          name: Age
-          type: date
-          priority: 0
-      schema:
-        openAPIV3Schema:
-          type: object
-          properties:
-            spec:
-              type: object
-              anyOf:
-                - required: ["http"]
-                - required: ["stream"]
-              properties:
-                ingressClassName:
-                  type: string
-                http:
-                  type: array
-                  minItems: 1
-                  items:
-                    type: object
-                    required: ["name", "match", "backends"]
-                    properties:
-                      name:
-                        type: string
-                        minLength: 1
-                      priority:
-                        type: integer
-                      timeout:
-                        type: object
-                        properties:
-                          connect:
-                            type: string
-                          send:
-                            type: string
-                          read:
-                            type: string
-                      match:
-                        type: object
-                        required:
-                          - paths
-                        properties:
-                          paths:
-                            type: array
-                            minItems: 1
-                            items:
-                              type: string
-                              pattern: "^/[a-zA-Z0-9\\-._~%!$&'()+,;=:@/]*\\*?$"
-                          hosts:
-                            type: array
-                            minItems: 1
-                            items:
-                              type: string
-                              pattern: "^\\*?[0-9a-zA-Z-._]+$"
-                          methods:
-                            type: array
-                            minItems: 1
-                            items:
-                              type: string
-                              enum:
-                                - "CONNECT"
-                                - "DELETE"
-                                - "GET"
-                                - "HEAD"
-                                - "OPTIONS"
-                                - "PATCH"
-                                - "POST"
-                                - "PUT"
-                                - "TRACE"
-                          remoteAddrs:
-                            type: array
-                            minItems: 1
-                            items:
-                              type: string
-                          exprs:
-                            type: array
-                            minItems: 1
-                            items:
-                              type: object
-                              properties:
-                                subject:
-                                  type: object
-                                  properties:
-                                    scope:
-                                      type: string
-                                      enum:
-                                        - "Cookie"
-                                        - "Header"
-                                        - "Path"
-                                        - "Query"
-                                    name:
-                                      type: string
-                                      minLength: 1
-                                  required:
-                                    - scope
-                                op:
-                                  type: string
-                                  enum:
-                                    - Equal
-                                    - NotEqual
-                                    - GreaterThan
-                                    - LessThan
-                                    - In
-                                    - NotIn
-                                    - RegexMatch
-                                    - RegexNotMatch
-                                    - RegexMatchCaseInsensitive
-                                    - RegexNotMatchCaseInsensitive
-                                value:
-                                  type: string
-                                set:
-                                  type: array
-                                  items:
-                                    type: string
-                              oneOf:
-                                - required: ["subject", "op", "value"]
-                                - required: ["subject", "op", "set"]
-                      websocket:
-                        type: boolean
-                      plugin_config_name:
-                        type: string
-                        minLength: 1
-                      backends:
-                        type: array
-                        minItems: 1
-                        items:
-                          type: object
-                          properties:
-                            serviceName:
-                              type: string
-                              minLength: 1
-                            servicePort:
-                              anyOf:
-                              - type: integer
-                              - type: string
-                              x-kubernetes-int-or-string: true
-                            resolveGranularity:
-                              type: string
-                              enum: ["endpoint", "service"]
-                            weight:
-                              type: integer
-                              minimum: 0
-                            subset:
-                              type: string
                         required:
                           - serviceName
                           - servicePort
